% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode}
\alias{encode}
\title{Encode data.frame of factors using One-Hot Encoding}
\usage{
encode(data, sep = ": ")
}
\arguments{
\item{data}{a data.frame where all components are factors.}

\item{sep}{a character defining the separator used for naming the columns of
encoded data.}
}
\value{
a matrix with encode factors.
}
\description{
Encode data.frame of factors using One-Hot Encoding
}
\author{
David Senhora Navega
}
