% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{icmr.control}
\alias{icmr.control}
\title{Inductive Confidence Machine for Regression Control}
\usage{
icmr.control(alpha = 0.05, process = T, interval = NULL)
}
\arguments{
\item{alpha}{a numeric defining the tolerance or confidence (1 - alpha) of
the predictive intervals of icmr model. Default is 0.05}

\item{process}{a logical stating if the data should be internally processed
by standardization.}

\item{interval}{a vector of the numeric values defining the domain of the
predictions generated by the icmr model. Default is NULL which means that
the interval is computed from the data.}
}
\value{
a list with the parameters given as arguments to the function call.
}
\description{
Inductive Confidence Machine for Regression Control
}
\author{
David Navega
}
