% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icmr.R
\name{icmr}
\alias{icmr}
\title{Inductive Confidence Machine for Regression}
\usage{
icmr(x, y, control = icmr.control())
}
\arguments{
\item{x}{a column-oriented matrix or a data.frame of numeric vector(s).}

\item{y}{a numeric vector}

\item{control}{a list of parameters created by \code{\link{icmr.control}}}
}
\value{
a icmr object
}
\description{
Reliable regression modelling using an Inductive Confidence Machine based on
Ridge Regression
}
\author{
David Senhora Navega
}
