% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icmr.R
\name{predict.icmr}
\alias{predict.icmr}
\title{Predict method for Inductive Confidence Machine for Regression}
\usage{
\method{predict}{icmr}(object, x, alpha = NULL, conformal = T, ...)
}
\arguments{
\item{object}{an icmr object}

\item{x}{a numeric vector or matrix with the same variables used to
create the icmr object}

\item{alpha}{level of error tolerance or confidence for predictive intervals
computed when conformal = T. Note confidence = 1 - alpha, with alpha assuming
a value between 0.01 and 0.5. Default value is obtained from the icmr object.}

\item{conformal}{a logical stating if confidence interval should be computed
using conformal prediction framework. See Details.}

\item{...}{...}
}
\description{
Predict method for Inductive Confidence Machine for Regression
}
\details{
...
}
\author{
David Senhora Navega
}
