% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icmr.R
\name{icmr.control}
\alias{icmr.control}
\title{Control List for Inductive Confidence Machine for Regression}
\usage{
icmr.control(alpha = 0.05, interval = NULL, delta = 0.25)
}
\arguments{
\item{alpha}{a numeric defining the tolerance or confidence (1 - alpha) of
the predictive intervals of icmr model. Default is 0.05}

\item{interval}{a vector of two numeric values defining the domain of the
predictions generated by the icmr model for a continuous output.
Default is NULL which means that the interval is computed from the data.}

\item{delta}{a numeric value between 0 and 1 that control the extrapolation
of the icmr model outside of the bounds of the values known for continuous
output. 0 means no extrapolation. This parameter affects both the
point estimates and the prediction intervals. Default is 0.25.}
}
\value{
a list with the parameters given as arguments to the function call.
}
\description{
Control List for Inductive Confidence Machine for Regression
}
\author{
David Navega
}
